requirements.txt
pyautogui>=0.9.54
Pillow>=10.0.0
numpy>=1.26.0
opencv-python-headless>=4.8.0
pytesseract>=0.3.10
mss>=9.0.1
sounddevice>=0.4.6
SoundFile>=0.12.1
SpeechRecognition>=3.10.0
pyaudio>=0.2.13
soundcard>=0.5.1
textblob>=0.17.1
spacy>=3.7.2
scikit-learn>=1.3.0
langdetect>=1.0.9
